Separate child test factories to its own file
